Refactoring Clip and OpenAi query encoders.
