+++ pyserini/search/faiss/_searcher.py
@@ -31,7 +31,7 @@ from transformers.file_utils import requires_backends
 
 from pyserini.encode import QueryEncoder, AutoQueryEncoder
 from pyserini.encode import AnceQueryEncoder, BprQueryEncoder, DprQueryEncoder, TctColBertQueryEncoder
-from pyserini.encode._clip import ClipEncoder
+from pyserini.encode.optional import ClipEncoder
 from pyserini.index import Document
 from pyserini.search.faiss._prf import PrfDenseSearchResult
 from pyserini.search.lucene import LuceneSearcher
