+++ requirements.txt
@@ -13,9 +13,9 @@ onnxruntime>=1.8.1
 openai>=1.0.0
 sentencepiece>=0.2
 tiktoken>=0.4.0
-# below are going to be "optional", eventually
-faiss>1.7.0
 flask>3.0
+# ones below are going to become optional dependencies eventually:
+faiss>1.7.0
 nmslib>=2.0.6
 lightgbm>=3.3.2
 spacy>=3.2.1
