+++ scripts/jobs.integrations-all.txt
@@ -1,4 +1,6 @@
-python -m unittest discover -s integrations/sparse > logs/log.sparse 2>&1
-python -m unittest discover -s integrations/dense > logs/log.dense 2>&1
-python -m unittest discover -s integrations/clprf > logs/log.clprf 2>&1
-python -m unittest discover -s integrations/papers > logs/log.papers 2>&1
+python -m unittest discover -s integrations/sparse > logs/log.core-sparse 2>&1
+python -m unittest discover -s integrations/clprf > logs/log.core-clprf 2>&1
+python -m unittest discover -s integrations/papers > logs/log.core-papers 2>&1
+python -m unittest discover -s integrations-optional/sparse > logs/log.optional-sparse 2>&1
+python -m unittest discover -s integrations-optional/dense > logs/log.optional-dense 2>&1
+python -m unittest discover -s integrations-optional/papers > logs/log.optional-papers 2>&1
